pyproject.toml
[tool.poetry]
name = "api_tests"
version = "0.1.0"
description = "Test example for ivi"
authors = ["Artyom Lagunkov"]

[tool.poetry.dependencies]
python = "^3.8"
allure-pytest = "^2.9.43"
requests = "^2.25.1"
colorlog = "^5.0.1"
pytruth = "^1.1.0"
pytest = "^6.2.4"
Faker = "^8.9.0"
pytest-rerunfailures = "^10.0"
pydantic = "^1.8.2"
dynaconf = "^3.1.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
